Add ioregionfd wire protocol
Adds constants and struct definitions to virt/kvm/ioregion.c.

This means the structs ioregionfd_cmd and ioregionfd_resp for commands
and responses, as well as the constants to break up the bits and bytes
in them.

And also adds ioregion for the mapped memory regions, and ioregionfd for
the read file descriptor.